Put shapes into different files
