+++ crates/epaint/src/shapes/mod.rs
@@ -1,6 +1,19 @@
+mod bezier;
+mod circle;
+mod ellipse;
+mod path;
+mod rect;
 mod shape;
+mod text;
 
-pub use shape::{
-    CircleShape, EllipseShape, PaintCallback, PaintCallbackInfo, PathShape, RectShape, Shape,
-    TextShape,
+pub use self::{
+    bezier::{CubicBezierShape, QuadraticBezierShape},
+    circle::CircleShape,
+    ellipse::EllipseShape,
+    path::PathShape,
+    rect::RectShape,
+    shape::PaintCallback,
+    shape::PaintCallbackInfo,
+    shape::Shape,
+    text::TextShape,
 };
