+++ crates/epaint/src/shapes/shape.rs
@@ -11,6 +11,8 @@ use emath::{pos2, Align2, Pos2, Rangef, Rect, TSTransform, Vec2};
 
 pub use crate::{CubicBezierShape, QuadraticBezierShape};
 
+use super::{CircleShape, EllipseShape, PathShape, RectShape, TextShape};
+
 /// A paint primitive such as a circle or a piece of text.
 /// Coordinates are all screen space points (not physical pixels).
 ///
@@ -460,418 +462,6 @@ impl Shape {
                 shape.rect = transform * shape.rect;
             }
         }
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// How to paint a circle.
-#[derive(Copy, Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct CircleShape {
-    pub center: Pos2,
-    pub radius: f32,
-    pub fill: Color32,
-    pub stroke: Stroke,
-}
-
-impl CircleShape {
-    #[inline]
-    pub fn filled(center: Pos2, radius: f32, fill_color: impl Into<Color32>) -> Self {
-        Self {
-            center,
-            radius,
-            fill: fill_color.into(),
-            stroke: Default::default(),
-        }
-    }
-
-    #[inline]
-    pub fn stroke(center: Pos2, radius: f32, stroke: impl Into<Stroke>) -> Self {
-        Self {
-            center,
-            radius,
-            fill: Default::default(),
-            stroke: stroke.into(),
-        }
-    }
-
-    /// The visual bounding rectangle (includes stroke width)
-    pub fn visual_bounding_rect(&self) -> Rect {
-        if self.fill == Color32::TRANSPARENT && self.stroke.is_empty() {
-            Rect::NOTHING
-        } else {
-            Rect::from_center_size(
-                self.center,
-                Vec2::splat(self.radius * 2.0 + self.stroke.width),
-            )
-        }
-    }
-}
-
-impl From<CircleShape> for Shape {
-    #[inline(always)]
-    fn from(shape: CircleShape) -> Self {
-        Self::Circle(shape)
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// How to paint an ellipse.
-#[derive(Copy, Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct EllipseShape {
-    pub center: Pos2,
-
-    /// Radius is the vector (a, b) where the width of the Ellipse is 2a and the height is 2b
-    pub radius: Vec2,
-    pub fill: Color32,
-    pub stroke: Stroke,
-}
-
-impl EllipseShape {
-    #[inline]
-    pub fn filled(center: Pos2, radius: Vec2, fill_color: impl Into<Color32>) -> Self {
-        Self {
-            center,
-            radius,
-            fill: fill_color.into(),
-            stroke: Default::default(),
-        }
-    }
-
-    #[inline]
-    pub fn stroke(center: Pos2, radius: Vec2, stroke: impl Into<Stroke>) -> Self {
-        Self {
-            center,
-            radius,
-            fill: Default::default(),
-            stroke: stroke.into(),
-        }
-    }
-
-    /// The visual bounding rectangle (includes stroke width)
-    pub fn visual_bounding_rect(&self) -> Rect {
-        if self.fill == Color32::TRANSPARENT && self.stroke.is_empty() {
-            Rect::NOTHING
-        } else {
-            Rect::from_center_size(
-                self.center,
-                self.radius * 2.0 + Vec2::splat(self.stroke.width),
-            )
-        }
-    }
-}
-
-impl From<EllipseShape> for Shape {
-    #[inline(always)]
-    fn from(shape: EllipseShape) -> Self {
-        Self::Ellipse(shape)
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// A path which can be stroked and/or filled (if closed).
-#[derive(Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct PathShape {
-    /// Filled paths should prefer clockwise order.
-    pub points: Vec<Pos2>,
-
-    /// If true, connect the first and last of the points together.
-    /// This is required if `fill != TRANSPARENT`.
-    pub closed: bool,
-
-    /// Fill is only supported for convex polygons.
-    pub fill: Color32,
-
-    /// Color and thickness of the line.
-    pub stroke: PathStroke,
-    // TODO(emilk): Add texture support either by supplying uv for each point,
-    // or by some transform from points to uv (e.g. a callback or a linear transform matrix).
-}
-
-impl PathShape {
-    /// A line through many points.
-    ///
-    /// Use [`Shape::line_segment`] instead if your line only connects two points.
-    #[inline]
-    pub fn line(points: Vec<Pos2>, stroke: impl Into<PathStroke>) -> Self {
-        Self {
-            points,
-            closed: false,
-            fill: Default::default(),
-            stroke: stroke.into(),
-        }
-    }
-
-    /// A line that closes back to the start point again.
-    #[inline]
-    pub fn closed_line(points: Vec<Pos2>, stroke: impl Into<PathStroke>) -> Self {
-        Self {
-            points,
-            closed: true,
-            fill: Default::default(),
-            stroke: stroke.into(),
-        }
-    }
-
-    /// A convex polygon with a fill and optional stroke.
-    ///
-    /// The most performant winding order is clockwise.
-    #[inline]
-    pub fn convex_polygon(
-        points: Vec<Pos2>,
-        fill: impl Into<Color32>,
-        stroke: impl Into<PathStroke>,
-    ) -> Self {
-        Self {
-            points,
-            closed: true,
-            fill: fill.into(),
-            stroke: stroke.into(),
-        }
-    }
-
-    /// The visual bounding rectangle (includes stroke width)
-    #[inline]
-    pub fn visual_bounding_rect(&self) -> Rect {
-        if self.fill == Color32::TRANSPARENT && self.stroke.is_empty() {
-            Rect::NOTHING
-        } else {
-            Rect::from_points(&self.points).expand(self.stroke.width / 2.0)
-        }
-    }
-}
-
-impl From<PathShape> for Shape {
-    #[inline(always)]
-    fn from(shape: PathShape) -> Self {
-        Self::Path(shape)
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// How to paint a rectangle.
-#[derive(Copy, Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct RectShape {
-    pub rect: Rect,
-
-    /// How rounded the corners are. Use `Rounding::ZERO` for no rounding.
-    pub rounding: Rounding,
-
-    /// How to fill the rectangle.
-    pub fill: Color32,
-
-    /// The thickness and color of the outline.
-    ///
-    /// The stroke extends _outside_ the edge of [`Self::rect`],
-    /// i.e. using [`crate::StrokeKind::Outside`].
-    ///
-    /// This means the [`Self::visual_bounding_rect`] is `rect.size() + 2.0 * stroke.width`.
-    pub stroke: Stroke,
-
-    /// If larger than zero, the edges of the rectangle
-    /// (for both fill and stroke) will be blurred.
-    ///
-    /// This can be used to produce shadows and glow effects.
-    ///
-    /// The blur is currently implemented using a simple linear blur in sRGBA gamma space.
-    pub blur_width: f32,
-
-    /// If the rect should be filled with a texture, which one?
-    ///
-    /// The texture is multiplied with [`Self::fill`].
-    pub fill_texture_id: TextureId,
-
-    /// What UV coordinates to use for the texture?
-    ///
-    /// To display a texture, set [`Self::fill_texture_id`],
-    /// and set this to `Rect::from_min_max(pos2(0.0, 0.0), pos2(1.0, 1.0))`.
-    ///
-    /// Use [`Rect::ZERO`] to turn off texturing.
-    pub uv: Rect,
-}
-
-impl RectShape {
-    /// The stroke extends _outside_ the [`Rect`].
-    #[inline]
-    pub fn new(
-        rect: Rect,
-        rounding: impl Into<Rounding>,
-        fill_color: impl Into<Color32>,
-        stroke: impl Into<Stroke>,
-    ) -> Self {
-        Self {
-            rect,
-            rounding: rounding.into(),
-            fill: fill_color.into(),
-            stroke: stroke.into(),
-            blur_width: 0.0,
-            fill_texture_id: Default::default(),
-            uv: Rect::ZERO,
-        }
-    }
-
-    #[inline]
-    pub fn filled(
-        rect: Rect,
-        rounding: impl Into<Rounding>,
-        fill_color: impl Into<Color32>,
-    ) -> Self {
-        Self {
-            rect,
-            rounding: rounding.into(),
-            fill: fill_color.into(),
-            stroke: Default::default(),
-            blur_width: 0.0,
-            fill_texture_id: Default::default(),
-            uv: Rect::ZERO,
-        }
-    }
-
-    /// The stroke extends _outside_ the [`Rect`].
-    #[inline]
-    pub fn stroke(rect: Rect, rounding: impl Into<Rounding>, stroke: impl Into<Stroke>) -> Self {
-        Self {
-            rect,
-            rounding: rounding.into(),
-            fill: Default::default(),
-            stroke: stroke.into(),
-            blur_width: 0.0,
-            fill_texture_id: Default::default(),
-            uv: Rect::ZERO,
-        }
-    }
-
-    /// If larger than zero, the edges of the rectangle
-    /// (for both fill and stroke) will be blurred.
-    ///
-    /// This can be used to produce shadows and glow effects.
-    ///
-    /// The blur is currently implemented using a simple linear blur in `sRGBA` gamma space.
-    #[inline]
-    pub fn with_blur_width(mut self, blur_width: f32) -> Self {
-        self.blur_width = blur_width;
-        self
-    }
-
-    /// The visual bounding rectangle (includes stroke width)
-    #[inline]
-    pub fn visual_bounding_rect(&self) -> Rect {
-        if self.fill == Color32::TRANSPARENT && self.stroke.is_empty() {
-            Rect::NOTHING
-        } else {
-            let Stroke { width, .. } = self.stroke; // Make sure we remember to update this if we change `stroke` to `PathStroke`
-            self.rect.expand(width + self.blur_width / 2.0)
-        }
-    }
-}
-
-impl From<RectShape> for Shape {
-    #[inline(always)]
-    fn from(shape: RectShape) -> Self {
-        Self::Rect(shape)
-    }
-}
-
-// ----------------------------------------------------------------------------
-
-/// How to paint some text on screen.
-///
-/// This needs to be recreated if `pixels_per_point` (dpi scale) changes.
-#[derive(Clone, Debug, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub struct TextShape {
-    /// Top left corner of the first character.
-    pub pos: Pos2,
-
-    /// The laid out text, from [`Fonts::layout_job`].
-    pub galley: Arc<Galley>,
-
-    /// Add this underline to the whole text.
-    /// You can also set an underline when creating the galley.
-    pub underline: Stroke,
-
-    /// Any [`Color32::PLACEHOLDER`] in the galley will be replaced by the given color.
-    /// Affects everything: backgrounds, glyphs, strikethrough, underline, etc.
-    pub fallback_color: Color32,
-
-    /// If set, the text color in the galley will be ignored and replaced
-    /// with the given color.
-    ///
-    /// This only affects the glyphs and will NOT replace background color nor strikethrough/underline color.
-    pub override_text_color: Option<Color32>,
-
-    /// If set, the text will be rendered with the given opacity in gamma space
-    /// Affects everything: backgrounds, glyphs, strikethrough, underline, etc.
-    pub opacity_factor: f32,
-
-    /// Rotate text by this many radians clockwise.
-    /// The pivot is `pos` (the upper left corner of the text).
-    pub angle: f32,
-}
-
-impl TextShape {
-    /// The given fallback color will be used for any uncolored part of the galley (using [`Color32::PLACEHOLDER`]).
-    ///
-    /// Any non-placeholder color in the galley takes precedence over this fallback color.
-    #[inline]
-    pub fn new(pos: Pos2, galley: Arc<Galley>, fallback_color: Color32) -> Self {
-        Self {
-            pos,
-            galley,
-            underline: Stroke::NONE,
-            fallback_color,
-            override_text_color: None,
-            opacity_factor: 1.0,
-            angle: 0.0,
-        }
-    }
-
-    /// The visual bounding rectangle
-    #[inline]
-    pub fn visual_bounding_rect(&self) -> Rect {
-        self.galley.mesh_bounds.translate(self.pos.to_vec2())
-    }
-
-    #[inline]
-    pub fn with_underline(mut self, underline: Stroke) -> Self {
-        self.underline = underline;
-        self
-    }
-
-    /// Use the given color for the text, regardless of what color is already in the galley.
-    #[inline]
-    pub fn with_override_text_color(mut self, override_text_color: Color32) -> Self {
-        self.override_text_color = Some(override_text_color);
-        self
-    }
-
-    /// Rotate text by this many radians clockwise.
-    /// The pivot is `pos` (the upper left corner of the text).
-    #[inline]
-    pub fn with_angle(mut self, angle: f32) -> Self {
-        self.angle = angle;
-        self
-    }
-
-    /// Render text with this opacity in gamma space
-    #[inline]
-    pub fn with_opacity_factor(mut self, opacity_factor: f32) -> Self {
-        self.opacity_factor = opacity_factor;
-        self
-    }
-}
-
-impl From<TextShape> for Shape {
-    #[inline(always)]
-    fn from(shape: TextShape) -> Self {
-        Self::Text(shape)
     }
 }
